-enabled deleting of items both from the master item list and from user wishlists -changed how prices are stored. now decimals. -require users to be logged in to see items

+++ apiapp/models.py
@@ -7,7 +7,7 @@ class Item(models.Model):
     name = models.CharField(max_length=60, default="Item")
     url = models.CharField(max_length=250, default="url")
     imageUrl = models.CharField(max_length=250, default="image")
-    price = models.IntegerField(default=0)
+    price = models.DecimalField(decimal_places=2, max_digits=12, default=0.00)
     description = models.CharField(max_length=250, default="description")
 
     def __str__(self):
